Applied most fixes
Changed directory and file distribution.
Used cleaner string concatenation and replace methods.

Still missing: one of the files isn't converting the initial header correctly even though it looks fine in markdown and is the same as the others.  const {readdirSync} was used but likely improperly as it ended up using more code.